item is done

+++ src/main/java/com/inventory/repository/ProductRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inventory.entity.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 	
-	Product findById(int id);
+	Product findById(long id);
 
 }
