+++ src/main/java/com/inventory/repository/PurchaseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inventory.entity.Purchase;
 
-public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
-	Purchase findById(int id);
+public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
+	Purchase findById(long id);
 }
